requirements.txt
psycopg2-binary
python-dotenv
streamlit
pandas
plotly
sqlalchemy
